grades en cours

+++ app/src/main/java/epitech/intratek/json/NsStat.java
@@ -4,12 +4,12 @@ import com.google.gson.annotations.SerializedName;
 
 public class NsStat
 {
-    public int active;
-    public int idle;
+    public float active;
+    public float idle;
     @SerializedName("out_active")
-    public int outActive;
+    public float outActive;
     @SerializedName("out_idle")
-    public int outIdle;
+    public float outIdle;
     @SerializedName("nslog_norm")
-    public int nsLogNorm;
+    public float nsLogNorm;
 }
